+++ polling-page-script-src/api/ApiService.ts
@@ -1,0 +1,10 @@
+/*
+ * SPDX-License-Identifier: CC0-1.0
+ */
+
+import type { Temperature } from "../utils/thermology/Temperature";
+
+export interface ApiService {
+
+	getTemperature(): PromiseLike<Temperature>;
+}