+++ CHANGELOG.md
@@ -11,11 +11,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [**Keep a Changelog v1.0.0**](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [**Semantic Versioning v2.0.0**](https://semver.org/spec/v2.0.0.html).
 
+## [v0.1.0-indev05] - 2023-09-10 ##
+
+[v0.1.0-indev05]: https://github.com/mfederczuk/usbtemp-server/releases/tag/v0.1.0-indev05
+
+### Added ###
+
+* The polling interval of the HTML polling page can be configured via the URL query parameter `pollingInterval`
+* The displayed decimal digits of the HTML page can be configured via the URL query parameters `minDecimalDigits` and
+  `maxDecimalDigits`
+
+### Changed ###
+
+* The default displayed decimal digits in the HTML page is now exactly 1
+
 ## [v0.1.0-indev04] - 2023-09-03 ##
 
-### Fixed ###
-
 [v0.1.0-indev04]: https://github.com/mfederczuk/usbtemp-server/releases/tag/v0.1.0-indev04
+
+### Fixed ###
 
 * Program would crash on startup if the virtual thermometers file does not exist